// src/api/gen/models/InfluencerProfile.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Media } from "./Media";
import type { SocialLink } from "./SocialLink";
export type InfluencerProfile = {
  userId: string;
  displayName?: string;
  bio?: string;
  avatar?: Media;
  socialLinks?: Array<SocialLink>;
  createdAt?: string;
  updatedAt?: string;
};