// src/api/gen/models/UpdateInfluencerProfileRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { SocialLink } from "./SocialLink";
export type UpdateInfluencerProfileRequest = {
  displayName?: string;
  bio?: string;
  avatarMediaId?: string;
  socialLinks?: Array<SocialLink>;
};